refactor(runtime): extract compaction cache load/write into AiSdkCompaction
Move the compact/synthesis-cache load and write family
(loadHistoryCompactBlocks, loadSynthesisCacheBlocks, writeSynthesisCacheBlocks,
writeHistoryCompactCheckpoint, writeHistoryCompactBlocks) and their exclusive
incrementRecord/mergeCountsInto helpers from AiSdkBackend into a new
AiSdkCompaction collaborator (packages/runtime/src/ai-sdk-compaction.ts).
AiSdkBackend instantiates the collaborator and delegates the call sites in
compactHistory, buildPriorMessages, and prepareContextBudgetPolicy to
this.compaction.*.

Turn-scoped priorRequestShape is no longer read inside the write methods; it is
threaded per call as requestShapeHashBefore, read at the stay-side call sites,
preserving the "previous turn's shape" semantics (cleanupAfterTurn still does
not clear it). The five methods change private -> public on the collaborator
because AiSdkBackend now calls them cross-class; the contract types stay in
ai-sdk-backend.ts (the collaborator type-imports AiSdkBackendInput).

Behavior-neutral pure move: runtime build + 2073 tests pass (0 fail, 7 skipped),
parity with the pre-split baseline. Part of #1084.
